requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paddlepaddle==2.5.2
paddleocr==2.7.3
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0 